Combat speed is now fractional; there is a buffer which saves up fractional combat speed for a speed boost. In stock this means that odd speed ships will only move faster every other turn.

+++ FrEee/Game/Objects/Space/Planet.cs
@@ -145,7 +145,7 @@ namespace FrEee.Game.Objects.Space
 		/// </summary>
 		public Colony Colony { get; set; }
 
-		public int CombatSpeed => 0;
+		public double CombatSpeed => 0;
 
 		public override ConstructionQueue ConstructionQueue
 		{
