+++ FrEee/Game/Objects/Vehicles/Troop.cs
@@ -18,7 +18,7 @@ namespace FrEee.Game.Objects.Vehicles
 			get { return AbilityTargets.Troop; }
 		}
 
-		public override int CombatSpeed => 0;
+		public override double CombatSpeed => 0;
 
 		public ICargoContainer Container
 		{
